update install, refactoring
