+++ Code/install.py
@@ -10,16 +10,7 @@ import formatter as f
 import milvus_manager as mm
 import neo4j_manager as n4m
 import text_processing as tp
-
-
-def normalize_embedding(emb):
-    norm = np.linalg.norm(emb)
-    if norm == 0:
-        return emb
-    else:
-        normalized_arr = emb / norm
-        return normalized_arr
-
+import normalize as nrm
 
 def main():
     warnings.filterwarnings("ignore", category=UserWarning, module="torch")
@@ -60,10 +51,10 @@ def main():
     for i, item in enumerate(tqdm(dataset, desc="Embedding Estratti", total=len(dataset))):
         intervention_id = item['audio_id']
         processed_embedding_audio = ap.get_audio_embedding(item['audio']['array'], model_audio)
-        processed_embedding_audio = normalize_embedding(processed_embedding_audio)
+        processed_embedding_audio = nrm.normalize_embedding(processed_embedding_audio)
         audio_embeddings.append({"intervention_id": intervention_id, "audio_embedding": processed_embedding_audio})
         processed_embedding_text = tp.get_text_embedding(item['normalized_text'], device, txt_tokenizer, txt_model)
-        processed_embedding_text = normalize_embedding(processed_embedding_text)
+        processed_embedding_text = nrm.normalize_embedding(processed_embedding_text)
         text_embeddings.append({"intervention_id": intervention_id, "text_embedding": processed_embedding_text})
     f.my_print(f'Estrazione Completata')
 
@@ -90,7 +81,9 @@ def main():
 
     f.my_print(f'Dati inseriti correttamente')
     f.my_print(f'Creazione Indici per le collezioni')
-    mm.create_indexes_for_collections(audio_collection, text_collection)
+    index_struct = mm.create_index_structure()
+    mm.create_index_collection(audio_collection, "audio_embedding", index_struct)
+    mm.create_index_collection(text_collection, "text_embedding", index_struct)
     f.my_print(f'Indici per le collezioni creati correttamente')
     mm.milvus_disconnect()
